Always provide metadata blob for hdr frames
If a surface is playing HDR10+ video and switches to content with a
different HDR standard, then the HDR10+ metadata must be cleared. In
particular, HLG is a standard that does not use HDR metadata, so
unfortunately sending nothing does not imply clearing the existing
metadata. So, we can fix this by sending an empty packet for per-frame
metadata.

Bug: 212496177
Test: builds
Change-Id: Ie3476c9d9bf32f789fdaad5a4fc3cc08e9adcce9
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Pranav Temkar <pranavtemkar@gmail.com>